Disable mult-stage build for API
The multistage build for API is giving me permisison errors whenever
the container is brought down and back up. This should rename the
mount point as "build" and then host the application from there.

It would be ideal to use this as a build container, then host the
application from another container, as then the build itself would
smaller.